+++ docs/spec/src/account/index.ts
@@ -1,10 +1,10 @@
-import { StepSpec } from '@jupiterone/integration-sdk-core';
+import { RelationshipClass, StepSpec } from '@jupiterone/integration-sdk-core';
 import { IntegrationConfig } from '../../../../src/config';
 
 export const accountSpec: StepSpec<IntegrationConfig>[] = [
   {
     /**
-     * ENDPOINT: n/a
+     * ENDPOINT: https://<portal>.travis-ci.com/user
      * PATTERN: Singleton
      */
     id: 'fetch-account',
@@ -12,11 +12,23 @@ export const accountSpec: StepSpec<IntegrationConfig>[] = [
     entities: [
       {
         resourceName: 'Account',
-        _type: 'acme_account',
+        _type: 'travisci_account',
         _class: ['Account'],
       },
+      {
+        resourceName: 'User',
+        _type: 'travisci_user',
+        _class: ['User'],
+      },
     ],
-    relationships: [],
+    relationships: [
+      {
+        _type: 'travisci_account_is_user',
+        sourceType: 'travisci_account',
+        _class: RelationshipClass.IS,
+        targetType: 'travisci_user',
+      },
+    ],
     dependsOn: [],
     implemented: true,
   },