+++ src/steps/account/converter.ts
@@ -3,23 +3,59 @@ import {
   Entity,
 } from '@jupiterone/integration-sdk-core';
 
+import { TravisCIUser } from '../../types';
 import { Entities } from '../constants';
 
-export function createAccountEntity(): Entity {
+export function getAccountKey(id: number): string {
+  return `travisci_account:${id.toString()}`;
+}
+
+export function createAccountEntity(account: {
+  id: number;
+  email: string;
+  hostname: string;
+  name: string;
+}): Entity {
   return createIntegrationEntity({
     entityData: {
-      source: {
-        id: 'acme-unique-account-id',
-        name: 'Example Co. Acme Account',
-      },
+      source: account,
       assign: {
-        _key: 'acme-unique-account-id',
+        _key: getAccountKey(account.id),
         _type: Entities.ACCOUNT._type,
         _class: Entities.ACCOUNT._class,
-        mfaEnabled: true,
-        // This is a custom property that is not a part of the data model class
-        // hierarchy. See: https://github.com/JupiterOne/data-model/blob/master/src/schemas/Account.json
-        manager: 'Manager Name',
+        id: account.id.toString(),
+        email: account.email,
+        hostname: account.hostname,
+        name: account.name,
+      },
+    },
+  });
+}
+
+/*
+   User & Account are the same entity. We can't iterate users.
+   Therefore we've added the user-related converter functions in here.
+   Alternative would be to have steps/users with only converter file.
+   Or we could have steps/users with its own step function but we'd
+   essentially be doing the same procedure just in two places which doesn't seems ideal.
+*/
+export function getUserKey(id: number): string {
+  return `travisci_user:${id.toString()}`;
+}
+
+export function createUserEntity(user: TravisCIUser): Entity {
+  return createIntegrationEntity({
+    entityData: {
+      source: user,
+      assign: {
+        _key: getUserKey(user.id),
+        _type: Entities.USER._type,
+        _class: Entities.USER._class,
+        id: user.id.toString(),
+        username: user.login,
+        email: user.email,
+        active: true,
+        name: user.name,
       },
     },
   });