+++ src/steps/account/index.ts
@@ -1,29 +1,57 @@
 import {
+  createDirectRelationship,
   IntegrationStep,
   IntegrationStepExecutionContext,
+  RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 
+import { createAPIClient } from '../../client';
 import { IntegrationConfig } from '../../config';
-import { Steps, Entities } from '../constants';
-import { createAccountEntity } from './converter';
+import { TravisCIUser } from '../../types';
+import {
+  ACCOUNT_ENTITY_KEY,
+  Steps,
+  Entities,
+  Relationships,
+} from '../constants';
+import { createAccountEntity, createUserEntity } from './converter';
 
-export const ACCOUNT_ENTITY_KEY = 'entity:account';
-
-export async function fetchAccountDetails({
+export async function fetchAccount({
+  instance,
   jobState,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
-  const accountEntity = await jobState.addEntity(createAccountEntity());
+  const apiClient = createAPIClient(instance.config);
 
+  const user: TravisCIUser = await apiClient.fetchUser();
+
+  const accountEntity = createAccountEntity({
+    email: user.email,
+    id: user.id,
+    hostname: instance.config.hostname,
+    name: user.name,
+  });
+  await jobState.addEntity(accountEntity);
   await jobState.setData(ACCOUNT_ENTITY_KEY, accountEntity);
+
+  const userEntity = createUserEntity(user);
+  await jobState.addEntity(userEntity);
+
+  await jobState.addRelationship(
+    createDirectRelationship({
+      _class: RelationshipClass.IS,
+      from: accountEntity,
+      to: userEntity,
+    }),
+  );
 }
 
 export const accountSteps: IntegrationStep<IntegrationConfig>[] = [
   {
     id: Steps.ACCOUNT,
     name: 'Fetch Account Details',
-    entities: [Entities.ACCOUNT],
-    relationships: [],
+    entities: [Entities.ACCOUNT, Entities.USER],
+    relationships: [Relationships.ACCOUNT_IS_USER],
     dependsOn: [],
-    executionHandler: fetchAccountDetails,
+    executionHandler: fetchAccount,
   },
 ];